Change from float to decimal for DiemCLO cal

+++ src/DTOs/BaiKiemTra/BaiKiemTraDTO.cs
@@ -6,6 +6,6 @@ public class BaiKiemTraDTO
 {
     public int Id { get; set; }
     public string Loai { get; set; } = string.Empty;
-    public float TrongSo { get; set; }
+    public decimal TrongSo { get; set; }
     public int LopHocPhanId { get; set; }
 }
