+++ src/DTOs/HocPhan/HocPhanDTO.cs
@@ -6,7 +6,7 @@ public class HocPhanDTO
 {
     public int Id { get; set; }
     public string Ten { get; set; } = string.Empty;
-    public float SoTinChi { get; set; }
+    public decimal SoTinChi { get; set; }
     public bool LaCotLoi { get; set; }
     public int? KhoaId { get; set; }
 }
